move clients to api
